# sn-external/src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OTUS Highload Architect",
    "version": "1.2.0"
  },
  "paths": {
...
    "/user/register": {
      "post": {
        "operationId": "registerUser",
        "description": "Регистрация нового пользователя",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная регистрация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Невалидные данные"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    },
    "/user/get/{id}": {
      "get": {
        "operationId": "getUser",
        "description": "Получение анкеты пользователя",
        "parameters": [
          {
            "name": "id",
            "schema": {
              "$ref": "#/components/schemas/UserId"
            },
            "required": true,
            "in": "path",
            "description": "Идентификатор пользователя"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное получение анкеты пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Невалидные данные"
          },
          "404": {
            "description": "Анкета не найдена"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    },
    "/user/search": {
      "get": {
        "operationId": "searchUser",
        "description": "Поиск анкет",
        "parameters": [
          {
            "name": "first_name",
            "schema": {
              "type": "string",
              "description": "Часть имени для поиска",
              "example": "Конст"
            },
            "in": "query",
            "required": true,
            "description": "Условие поиска по имени"
          },
          {
            "name": "second_name",
            "schema": {
              "type": "string",
              "description": "Часть фамилии для поиска",
              "example": "Оси"
            },
            "in": "query",
            "required": true,
            "description": "Условие поиска по фамилии"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешные поиск пользователя",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Невалидные данные"
          },
          "500": {
            "$ref": "#/components/responses/5xx"
          },
          "503": {
            "$ref": "#/components/responses/5xx"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "5xx": {
        "description": "Ошибка сервера",
        "headers": {
          "Retry-After": {
            "description": "Время, через которое еще раз нужно сделать запрос",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Описание ошибки"
                },
                "request_id": {
                  "type": "string",
                  "description": "Идентификатор запроса. Предназначен для более быстрого поиска проблем."
                },
                "code": {
                  "type": "integer",
                  "description": "Код ошибки. Предназначен для классификации проблем и более быстрого решения проблем."
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "BirthDate": {
        "type": "string",
        "description": "Дата рождения",
        "format": "date",
        "example": "2017-02-01"
      },
      "UserId": {
        "type": "string",
        "description": "Идентификатор пользователя"
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "example": "Имя"
          },
          "second_name": {
            "type": "string",
            "example": "Фамилия"
          },
          "birthdate": {
            "$ref": "#/components/schemas/BirthDate"
          },
          "biography": {
            "type": "string",
            "example": "Хобби, интересы и т.п."
          },
          "city": {
            "type": "string",
            "example": "Москва"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "Секретная строка"
          }
        },
        "required": [
          "first_name", "second_name", "password"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "$ref": "#/components/schemas/UserId"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "$ref": "#/components/schemas/UserId"
          },
          "first_name": {
            "type": "string",
            "example": "Имя",
            "description": "Имя"
          },
          "second_name": {
            "type": "string",
            "example": "Фамилия",
            "description": "Фамилия"
          },
          "birthdate": {
            "$ref": "#/components/schemas/BirthDate"
          },
          "biography": {
            "type": "string",
            "example": "Хобби, интересы и т.п.",
            "description": "Интересы"
          },
          "city": {
            "type": "string",
            "example": "Москва",
            "description": "Город"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Авторизация по токену, который был получен в методе /login"
      }
    }
  }
}